+++ brewtils/test/fixtures.py
@@ -7,6 +7,7 @@ import pytest
 import pytz
 
 from brewtils.models import (
+    AliasUserMap,
     Choices,
     Command,
     Connection,
@@ -17,22 +18,24 @@ from brewtils.models import (
     Instance,
     IntervalTrigger,
     Job,
-    LegacyRole,
     LoggingConfig,
     Operation,
     Parameter,
     PatchOperation,
-    Principal,
     Queue,
     Replication,
     Request,
     RequestFile,
     RequestTemplate,
     Resolvable,
+    Role,
     Runner,
-    System,
     Subscriber,
+    System,
     Topic,
+    UpstreamRole,
+    User,
+    UserToken,
 )
 
 
@@ -263,6 +266,7 @@ def system_dict(instance_dict, command_dict, command_dict_2, system_id):
         "local": True,
         "template": "<html>template</html>",
         "groups": ["GroupB", "GroupA"],
+        "prefix_topic": "custom_topic",
     }
 
 
@@ -531,38 +535,107 @@ def bg_queue(queue_dict):
 
 
 @pytest.fixture
-def principal_dict(legacy_role_dict):
+def alias_user_map_dict():
     return {
-        "id": "58542eb571afd47ead90d24f",
-        "username": "admin",
-        "roles": [legacy_role_dict],
-        "permissions": ["bg-all"],
-        "preferences": {"theme": "dark"},
-        "metadata": {"foo": "bar"},
+        "target_garden": "test",
+        "username": "user",
     }
 
 
 @pytest.fixture
-def bg_principal(principal_dict, bg_role):
-    dict_copy = copy.deepcopy(principal_dict)
-    dict_copy["roles"] = [bg_role]
-    return Principal(**dict_copy)
+def bg_alias_user_map(alias_user_map_dict):
+    return AliasUserMap(**alias_user_map_dict)
 
 
 @pytest.fixture
-def legacy_role_dict():
+def user_token_dict(user_dict, ts_epoch):
     return {
-        "id": "58542eb571afd47ead90d26f",
-        "name": "bg-admin",
-        "description": "The admin role",
-        "permissions": ["bg-all"],
+        "id": "1",
+        "uuid": "11111111-2222-4444-5555-66666666666",
+        "issued_at": ts_epoch,
+        "expires_at": ts_epoch,
+        "username": "USERNAME",
     }
 
 
 @pytest.fixture
-def bg_role(legacy_role_dict):
-    dict_copy = copy.deepcopy(legacy_role_dict)
-    return LegacyRole(**dict_copy)
+def bg_user_token(user_token_dict, ts_dt):
+    dict_copy = copy.deepcopy(user_token_dict)
+    dict_copy["issued_at"] = ts_dt
+    dict_copy["expires_at"] = ts_dt
+    return UserToken(**dict_copy)
+
+
+@pytest.fixture
+def role_dict():
+    return {
+        "permission": "PLUGIN_ADMIN",
+        "description": "PLUGIN ADMIN ROLE",
+        "id": "1",
+        "name": "PLUGIN_ADMIN_ROLE",
+        "scope_gardens": ["FOO"],
+        "scope_namespaces": [],
+        "scope_systems": [],
+        "scope_instances": [],
+        "scope_versions": [],
+        "scope_commands": [],
+        "protected": False,
+        "file_generated": False,
+    }
+
+
+@pytest.fixture
+def bg_role(role_dict):
+    return Role(**role_dict)
+
+
+@pytest.fixture
+def upstream_role_dict():
+    return {
+        "permission": "PLUGIN_ADMIN",
+        "description": "PLUGIN ADMIN ROLE",
+        "id": "1",
+        "name": "PLUGIN_ADMIN_ROLE",
+        "scope_gardens": ["FOO"],
+        "scope_namespaces": [],
+        "scope_systems": [],
+        "scope_instances": [],
+        "scope_versions": [],
+        "scope_commands": [],
+        "protected": False,
+        "file_generated": False,
+    }
+
+
+@pytest.fixture
+def bg_upstream_role(upstream_role_dict):
+    return UpstreamRole(**upstream_role_dict)
+
+
+@pytest.fixture
+def user_dict(role_dict, upstream_role_dict, alias_user_map_dict):
+    return {
+        "id": "1",
+        "username": "USERNAME",
+        "password": "HASH",
+        "roles": ["PLUGIN_ADMIN_ROLE"],
+        "local_roles": [role_dict],
+        "upstream_roles": [upstream_role_dict],
+        "user_alias_mapping": [alias_user_map_dict],
+        "is_remote": False,
+        "metadata": {},
+        "protected": False,
+        "file_generated": False,
+    }
+
+
+@pytest.fixture
+def bg_user(user_dict, bg_role, bg_upstream_role, bg_alias_user_map):
+    dict_copy = copy.deepcopy(user_dict)
+    dict_copy["upstream_roles"] = [bg_upstream_role]
+    dict_copy["local_roles"] = [bg_role]
+    dict_copy["user_alias_mapping"] = [bg_alias_user_map]
+    return User(**dict_copy)
 
 
 @pytest.fixture
@@ -646,7 +719,6 @@ def job_dict_for_import(job_dict):
     """A job dict but some keys and values are missing."""
     dict_copy = copy.deepcopy(job_dict)
     for field in [
-        "id",
         "next_run_time",
         "success_count",
         "error_count",
@@ -836,6 +908,8 @@ def garden_dict(ts_epoch, system_dict, connection_dict, connection_publishing_di
         "has_parent": False,
         "children": [],
         "metadata": {},
+        "default_user": None,
+        "shared_users": True,
     }
 
 
@@ -935,6 +1009,7 @@ def subscriber_dict():
         "version": "1.0.0",
         "instance": "inst",
         "command": "run",
+        "subscriber_type": "DYNAMIC",
     }